+++ client/src/pages/UserSpace/ArchivePage/Content.js
@@ -1,0 +1,13 @@
+import React, { useEffect } from 'react'
+import { useHttp } from 'hooks'
+import st from 'styles/UserSpace/ArchivePage/content.module.css'
+
+export const Content = () => {
+  const { request, loading } = useHttp()
+  
+  return (
+    <div className={st.cards_area}>
+
+    </div>
+  )
+}
